+++ public/api/me.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+require __DIR__ . '/_bootstrap.php';
+
+apiRequireMethod('GET');
+
+apiRespond($services['authController']->me());